Changed this repository to store tests
atm it is using robotium to do basic tests, as a lot of our buttons are
missing it isn't testing much.  There's something weird with clicking on
buttons which change activities, I had to use sleep() to get it to work.
Maybe just something to do with how slow the emulator is.